feat: enable push to multiple registries
