+++ .github/workflows/continuous-delivery.yaml
@@ -1,9 +1,6 @@
 name: Continuous Delivery
 
-on:
-  push:
-    branches:
-      - master
+on: push
 
 jobs:
   release:
@@ -39,6 +36,7 @@ jobs:
         env:
           # This must be used as GitHub Actions token does not support pushing
           # to protected branches.
+          DRY_RUN: true
           GITHUB_TOKEN: ${{ secrets.GITHUB_TOKEN_DOTTBOTT }}
           NPM_TOKEN: ${{ secrets.GITHUB_TOKEN_DOTTBOTT }}
         with:
